requirements.txt
fastapi>=0.95.2
uvicorn[standard]>=0.22.0
requests>=2.31.0
pydantic>=1.10.12,<2.0.0
python-dotenv>=1.0.0

# Voice and Speech
SpeechRecognition>=3.10.0
pyttsx3>=2.90
elevenlabs>=0.2.27
pygame>=2.5.2

# Research and Web Scraping
beautifulsoup4>=4.12.0
googlesearch-python>=1.2.3
aiohttp>=3.9.0

# Platform Integrations
google-auth>=2.27.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.111.0
slack-sdk>=3.26.0
notion-client>=2.2.1
pytest
keyring==24.3.0
cryptography==41.0.0